requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-PyMongo==2.3.0
pymongo==4.5.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy>=1.24.0
python-dotenv==1.0.0
APScheduler==3.10.4
dnspython==2.4.2